class City was deleted due to uncessesary and united as stati method to HH class. test_for get_city_id method is done and moved to test_hh file. new class module is created for json_management class.
