+++ src/json_management.py
@@ -1,0 +1,13 @@
+from vacancy import Vacancy
+from abc import ABC, abstractmethod
+
+
+class ManageJSON(ABC):
+    """Класс-шаблон для определения методов, а так же атрибутов"""
+    pass
+
+
+class JSONManagement(Vacancy):
+    """Класс для управления данными, для записи, сортировки, обработки и """
+    pass
+
