+++ src/links_constants.py
@@ -1,5 +1,5 @@
 """модуль с хранениями констант и ссылок"""
 
 hh_API_vacancies = 'https://api.hh.ru/vacancies'
-hh_API_cities = 'https://api.hh.ru/areas'
+HH_API_CITIES = 'https://api.hh.ru/areas'
 HEADERS = {'User-Agent': 'HH-User-Agent'}
